packages/cleafy-plugin/android/src/main/kotlin/com/cleafy/cleafyPlugin/methods/AppDeviceIdFlutterMethod.kt
package com.cleafy.cleafyPlugin.methods

import com.cleafy.cleafyPlugin.interop.CleafyProxy
import io.flutter.embedding.engine.plugins.FlutterPlugin
import io.flutter.plugin.common.MethodCall
import io.flutter.plugin.common.MethodChannel

internal class AppDeviceIdFlutterMethod(private val cleafy: CleafyProxy) : FlutterMethod {

    override fun invoke(
        flutterPluginBinding: FlutterPlugin.FlutterPluginBinding,
        call: MethodCall,
        result: MethodChannel.Result
    ) {
        val deviceId = call.arguments<String>() ?: run {
            result.error("", "App device ID not found", null)
            return
        }
        cleafy.setAppDeviceId(deviceId)
        result.success("New app device ID set!")
    }
}
